+++ NEWS.md
@@ -5,6 +5,7 @@
 * `caret2rsample` and `rsample2caret` can convert `rset` objects to those used by `caret::trainControl` and vice-versa. 
 * A function called `form_pred` can be used to determine the original names of the predictors in a formula or `terms` object. 
 * A vignette and a function (`prepper`) were included to facilitate using the `recipes` with `rsample`.
+* A `gather` method was added for `rset` objects.
 
 # `rsample` 0.0.1